Changed how error estimates are formatted.
Introduced a new default style of printing the error estimates, using parenthesis notation instead of plus-minus. The new style is more compact, and hence the default since space is at a premium. At the same time I introduced the argument 'error_style' which chooses the preferred style. Default is 'parenthesis', other options are 'plusminus' (for the old plus-minus-style) and None for only showing the median without error estimates.

In connection to this I also made sure that integer estimates can be rounded and handled correctly.