Improved automatic figsize.
The automatic figsize now moves in jumps: if ndim is less than 8, the figsize is (base, base) where base=4.8.
If 8 <= ndim < 16, the figsize is (2*base, 2*base). And so on.
With these figsizes a plot with ndim < 8 should fit in a single column of a paper, while 8 <= ndim < 16 should fit a double column. Larger plots will probably never fit in any document, but the scaling seems reasonable.